+++ server/src/index.js
@@ -4,12 +4,14 @@ import { runMigrations, bootstrapAdmin, seedChangeTypesFromConfig } from './db/m
 import { createApp } from './app.js';
 import { startScheduler, stopScheduler } from './services/digestScheduler.js';
 import { startEmailPoller, stopEmailPoller } from './services/emailPoller.js';
+import { startRecurringScheduler, stopRecurringScheduler } from './services/recurringScheduler.js';
 
 runMigrations();
 bootstrapAdmin();
 seedChangeTypesFromConfig();
 startScheduler();
 startEmailPoller();
+startRecurringScheduler();
 
 const app = createApp();
 const server = app.listen(config.port, () => {
@@ -20,6 +22,7 @@ const shutdown = (sig) => {
   logger.info({ sig }, 'shutting down');
   stopScheduler();
   stopEmailPoller();
+  stopRecurringScheduler();
   server.close(() => process.exit(0));
 };
 process.on('SIGINT', shutdown);